Separate out `domain_extrinsics_builder` module in subspace-fraud-proof
